+++ Source/FDAudioContext.swift
@@ -1,0 +1,60 @@
+//
+// Copyright 2013 - 2017, William Entriken and the FDWaveformView contributors.
+//
+import UIKit
+import AVFoundation
+
+/// Holds audio information used for building waveforms
+final class FDAudioContext {
+    
+    /// The audio asset URL used to load the context
+    public let audioURL: URL
+    
+    /// Total number of samples in loaded asset
+    public let totalSamples: Int
+    
+    /// Loaded asset
+    public let asset: AVAsset
+    
+    // Loaded assetTrack
+    public let assetTrack: AVAssetTrack
+    
+    private init(audioURL: URL, totalSamples: Int, asset: AVAsset, assetTrack: AVAssetTrack) {
+        self.audioURL = audioURL
+        self.totalSamples = totalSamples
+        self.asset = asset
+        self.assetTrack = assetTrack
+    }
+    
+    public static func load(fromAudioURL audioURL: URL, completionHandler: @escaping (_ audioContext: FDAudioContext?) -> ()) {
+        let asset = AVURLAsset(url: audioURL, options: [AVURLAssetPreferPreciseDurationAndTimingKey: NSNumber(value: true as Bool)])
+        
+        guard let assetTrack = asset.tracks(withMediaType: AVMediaTypeAudio).first else {
+            NSLog("FDWaveformView failed to load AVAssetTrack")
+            completionHandler(nil)
+            return
+        }
+        
+        asset.loadValuesAsynchronously(forKeys: ["duration"]) {
+            var error: NSError?
+            let status = asset.statusOfValue(forKey: "duration", error: &error)
+            switch status {
+            case .loaded:
+                guard
+                    let audioFormatDesc = assetTrack.formatDescriptions.first,
+                    let asbd = CMAudioFormatDescriptionGetStreamBasicDescription(audioFormatDesc as! CMAudioFormatDescription) // TODO: Can this be safer?
+                    else { break }
+                
+                let totalSamples = Int((asbd.pointee.mSampleRate) * Float64(asset.duration.value) / Float64(asset.duration.timescale))
+                let audioContext = FDAudioContext(audioURL: audioURL, totalSamples: totalSamples, asset: asset, assetTrack: assetTrack)
+                completionHandler(audioContext)
+                return
+                
+            case .failed, .cancelled, .loading, .unknown:
+                print("FDWaveformView could not load asset: \(error?.localizedDescription ?? "Unknown error")")
+            }
+            
+            completionHandler(nil)
+        }
+    }
+}
